requirements.txt
pyTelegramBotAPI==4.14.0
requests==2.31.0
qrcode==7.4.2
Pillow==10.1.0
python-dotenv==1.0.0
